+++ tool/html_generator/lib/src/common/header_html.dart
@@ -1,27 +1,35 @@
 /// HTML template of the `<header>` tag.
 const headerHtml = '''
-    <meta name="keywords" content="Flutter Complete Reference, Flutter Book, Dart Book">
+    <meta name="keywords" content="Flutter Complete Reference 2, Flutter Book, Dart Book">
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <meta name="description" content="The definitive reference book for Dart 3.0 and Flutter.">
     <meta name="author" content="Alberto Miola">
     <meta name="google-site-verification" content="o4mSLG8Q94yhenxMlyMWtIJ97Bvws6smxCmdDHtjNjg">
+    <meta name="apple-mobile-web-app-title" content="Flutter Complete Reference 2">
 
     <meta property="og:title" content="Flutter Complete Reference 2">
     <meta property="og:site_name" content="Flutter Complete Reference 2">
     <meta property="og:type" content="book">
     <meta property="og:url" content="https://fluttercompletereference.com/">
-    <meta property="og:image" content="https://fluttercompletereference.com/images/twitter_logo.png">
+    <meta property="og:image" content="https://fluttercompletereference.com/images/logo.png">
     <meta property="og:description" content="The definitive reference book for Dart 3.0 and Flutter.">
 
     <meta name="twitter:card" content="summary">
     <meta name="twitter:site" content="@flutter_book">
     <meta name="twitter:title" content="Flutter Complete Reference 2">
     <meta name="twitter:description" content="The definitive reference book for Dart 3.0 and Flutter.">
-    <meta name="twitter:image" content="https://fluttercompletereference.com/images/png_logo.png">
+    <meta name="twitter:image" content="https://fluttercompletereference.com/images/logo.png">
 
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700&display=swap" rel="stylesheet">
-    <link href="styles/style_min.css" rel="stylesheet">
+    <link href="/styles/style_min.css" rel="stylesheet">
+    
+    <link rel="icon" type="image/png" sizes="16x16" href="/images/favicon_16.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/images/favicon_32.png">
+    <link rel="icon" type="image/png" sizes="96x96" href="/images/favicon_96.png">
+    <link rel="icon" type="image/png" sizes="192x192" href="/images/favicon_192.png">
+    
+    <link rel="manifest" href="/manifest.json">
 ''';