+++ tool/html_generator/lib/src/common/navigation_bar_desktop.dart
@@ -1,9 +1,9 @@
-/// HTML template of the main navigation bar.
+/// HTML template of the desktop navigation bar.
 const navigationBarDesktopHtml = '''
     <header>
         <div class="logo_title">
             <img src="/images/svg/logo.svg" alt="Website logo" height="30">
-            <p>Complete Reference <span>2</span></p>
+            <p>Complete Reference <span>2.0</span></p>
         </div>
 
         <nav>