+++ tool/html_generator/lib/src/generators/overview_page_generator.dart
@@ -18,11 +18,11 @@ final class OverviewPageGenerator extends BasePageGenerator {
 const _overviewHtmlPage = '''
     <div class="overview-book-part">
         <h1>Part 1</h1>
-        <h2>The first part is a comprehensive coverage of the Dart 3.x language and is made up of 7 chapters:</h2>
+        <h2>The first part is about the Dart 3 programming language. It has eight chapters:</h2>
 
         <div class="chapters-container">
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">2</span> <span class="span-chapter-title">- Variables and data types</span></p>
+                <p>Chapter <span class="span-chapter-number">2</span> <span class="span-chapter-title">- Variables and types</span></p>
                 <ul>
                     <li>Variables</li>
                     <li>Data types</li>
@@ -46,7 +46,7 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">5</span> <span class="span-chapter-title">- Inheritance and exceptions</span></p>
+                <p>Chapter <span class="span-chapter-number">5</span> <span class="span-chapter-title">- Inheritance, core classes and exceptions</span></p>
                 <ul>
                     <li>Inheritance</li>
                     <li>The Object class</li>
@@ -62,7 +62,15 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">7</span> <span class="span-chapter-title">- Futures, Streams and Isolates</span></p>
+                <p>Chapter <span class="span-chapter-number">7</span> <span class="span-chapter-title">- Advanced language features</span></p>
+                <ul>
+                    <li>Class modifiers</li>
+                    <li>Patterns</li>
+                    <li>Exhaustiveness checking</li>
+                </ul>
+            </div>
+            <div class="chapter-entry">
+                <p>Chapter <span class="span-chapter-number">8</span> <span class="span-chapter-title">- Futures, Streams and Isolates</span></p>
                 <ul>
                     <li>Futures</li>
                     <li>Streams</li>
@@ -70,7 +78,7 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">8</span> <span class="span-chapter-title">- API overview and tests</span></p>
+                <p>Chapter <span class="span-chapter-number">9</span> <span class="span-chapter-title">- API overview and tests</span></p>
                 <ul>
                     <li>Dart core libraries</li>
                     <li>Testing Dart code</li>
@@ -84,11 +92,11 @@ const _overviewHtmlPage = '''
 
     <div class="overview-book-part">
         <h1>Part 2</h1>
-        <h2>The second part is a comprehensive coverage of the Flutter 4.x framework and is made up of 11 chapters:</h2>
+        <h2>The second part is about the Flutter 3.10 framework. It has ten chapters:</h2>
 
         <div class="chapters-container">
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">9</span> <span class="span-chapter-title">- Flutter, widgets and trees</span></p>
+                <p>Chapter <span class="span-chapter-number">10</span> <span class="span-chapter-title">- Flutter, widgets and trees</span></p>
                 <ul>
                     <li>Flutter overview</li>
                     <li>Widgets</li>
@@ -96,15 +104,15 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">10</span> <span class="span-chapter-title">- Material, Cupertino and CustomPaint</span></p>
+                <p>Chapter <span class="span-chapter-number">11</span> <span class="span-chapter-title">- Material, Cupertino and CustomPaint</span></p>
                 <ul>
                     <li>Basic widgets</li>
                     <li>Material and Cupertino library</li>
-                    <li>CustomPaint and CustomPainter</li>
+                    <li>Custom painters</li>
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">11</span> <span class="span-chapter-title">- State management</span></p>
+                <p>Chapter <span class="span-chapter-number">12</span> <span class="span-chapter-title">- State management</span></p>
                 <ul>
                     <li>Using setState</li>
                     <li>Using InheritedWidget</li>
@@ -112,30 +120,31 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">12</span> <span class="span-chapter-title">- Routes and navigation</span></p>
+                <p>Chapter <span class="span-chapter-number">13</span> <span class="span-chapter-title">- Routes and navigation</span></p>
                 <ul>
                     <li>Imperative and declarative navigation</li>
                     <li>The go_router package</li>
-                    <li>Comparison</li>
+                    <li>Deep linking and comparisons</li>
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">13</span> <span class="span-chapter-title">- Layout and responsiveness</span></p>
+                <p>Chapter <span class="span-chapter-number">14</span> <span class="span-chapter-title">- Layouts and responsiveness</span></p>
                 <ul>
                     <li>Layouts in Flutter</li>
                     <li>Building responsive layouts</li>
-                    <li>Scrolling widgets</li>
+                    <li>Scrolling widgets and Slivers</li>
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">14</span> <span class="span-chapter-title">- Internationalization and accessibility</span></p>
+                <p>Chapter <span class="span-chapter-number">15</span> <span class="span-chapter-title">- Internationalization and accessibility</span></p>
                 <ul>
                     <li>Internationalization</li>
                     <li>Accessibility</li>
+                    <li>The intl package</li>
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">15</span> <span class="span-chapter-title">- Assets and images</span></p>
+                <p>Chapter <span class="span-chapter-number">16</span> <span class="span-chapter-title">- Assets and images</span></p>
                 <ul>
                     <li>Defining assets</li>
                     <li>Images in Flutter</li>
@@ -143,7 +152,7 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">16</span> <span class="span-chapter-title">- Animations</span></p>
+                <p>Chapter <span class="span-chapter-number">17</span> <span class="span-chapter-title">- Animations</span></p>
                 <ul>
                     <li>Implicit animations</li>
                     <li>Explicit animations</li>
@@ -151,7 +160,7 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">17</span> <span class="span-chapter-title">- Forms and gestures</span></p>
+                <p>Chapter <span class="span-chapter-number">18</span> <span class="span-chapter-title">- Forms and gestures</span></p>
                 <ul>
                     <li>Form input widgets</li>
                     <li>Handling forms</li>
@@ -159,29 +168,19 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">18</span> <span class="span-chapter-title">- Testing Flutter applications</span></p>
+                <p>Chapter <span class="span-chapter-number">19</span> <span class="span-chapter-title">- Testing Flutter applications</span></p>
                 <ul>
-                    <li>Introduction to widget testing</li>
-                    <li>Widget testing strategies</li>
+                    <li>Widget testing</li>
+                    <li>Golden tests</li>
                     <li>Integration tests</li>
                 </ul>
-            </div>
-            <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">19</span> <span class="span-chapter-title">- Performance and profiling</span></p>
-                <ul>
-                    <li>Working with DevTools</li>
-                    <li>Performance best practices</li>
-                    <li>Profiling Flutter apps</li>
-                </ul>
-            </div>
-            <div class="chapter-entry-placeholder">
             </div>
         </div>
     </div>
 
     <div class="overview-book-part">
         <h1>Part 3</h1>
-        <h2>The third part of the book is about Dart and Flutter example applications. It has six chapters:</h2>
+        <h2>The third part is about the Dart and Flutter ecosystems. It has four chapters:</h2>
 
         <div class="chapters-container">
             <div class="chapter-entry">
@@ -201,33 +200,41 @@ const _overviewHtmlPage = '''
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">22</span> <span class="span-chapter-title">- Networking and caching strategies</span></p>
+                <p>Chapter <span class="span-chapter-number">22</span> <span class="span-chapter-title">- HTTP servers and low-level HTML</span></p>
                 <ul>
-                    <li>Using the http package</li>
-                    <li>Networking in Flutter</li>
-                    <li>Data caching in Flutter</li>
+                    <li>Creating HTTP servers</li>
+                    <li>The shelf package</li>
+                    <li>HTML manipulation with dart:html</li>
                 </ul>
             </div>
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">23</span> <span class="span-chapter-title">- Creating servers with Dart</span></p>
-                <ul>
-                    <li>Creating servers with dart:io</li>
-                    <li>Creating servers with shelf</li>
-                    <li>Creating and maintaining a server</li>
-                </ul>
-            </div>
-            <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">24</span> <span class="span-chapter-title">- Platform interactions</span></p>
+                <p>Chapter <span class="span-chapter-number">23</span> <span class="span-chapter-title">- Platform interactions</span></p>
                 <ul>
                     <li>Platform specific features</li>
                     <li>Creating plugin packages</li>
                     <li>Creating FFI plugin packages</li>
                 </ul>
             </div>
+        </div>
+    </div>
+
+    <div class="overview-book-part extra-space-overview-book">
+        <h1>Extra</h1>
+        <h2>There are some extra contents on the book.</h2>
+
+        <div class="chapters-container">
             <div class="chapter-entry">
-                <p>Chapter <span class="span-chapter-number">25</span> <span class="span-chapter-title">- Firebase and Flutter</span></p>
+                <p>Chapter <span class="span-chapter-number">1</span> <span class="span-chapter-title">- Introduction</span></p>
                 <ul>
-                    <li>TBA</li>
+                    <li>Dart introduction</li>
+                    <li>Flutter introduction</li>
+                </ul>
+            </div>
+            <div class="chapter-entry">
+                <p>Appendix</p>
+                <ul>
+                    <li>DevTools</li>
+                    <li>Performance best practices</li>
                 </ul>
             </div>
         </div>