+++ tool/html_generator/test/common/navigator_bar_desktop_test.dart
@@ -10,7 +10,7 @@ void main() {
     <header>
         <div class="logo_title">
             <img src="/images/svg/logo.svg" alt="Website logo" height="30">
-            <p>Complete Reference <span>2</span></p>
+            <p>Complete Reference <span>2.0</span></p>
         </div>
 
         <nav>