+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.0.0
 
-- Initial version of the HTTP server, published at https://fluttercompletereference.com/.
+- Initial version of the HTTP server, published at https://fluttercompletereference.com.